^r Extract mutating subcommands from authpolicy manage.go, move-only (gofmt, vet, build, package tests pass; declaration set verified identical; internal restructure, no behavior change)
Step 1 of 3 for the authpolicy god-file split: commandInit, commandSet,
commandUnset, and their helpers move verbatim to manage_mutate.go.
